requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-celery-results==1.0.1
djangorestframework==3.8.2
et-xmlfile==1.0.1
flickrapi==2.4.0
idna==2.7
jdcal==1.4
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
oauthlib==2.1.0
openpyxl==2.5.10
pytz==2018.4
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
six==1.11.0
UNKNOWN==0.0.0
urllib3==1.23
vine==1.1.4
virtualenv==16.1.0
